requirements.txt
flask==2.2.3
line-bot-sdk==2.4.2
python-dotenv==1.0.0
pyngrok==6.0.0
